+++ _goldens/test_files/opaque_token.dart
@@ -14,3 +14,35 @@ class HasOpaqueTokens {
 
   HasOpaqueTokens(@Inject(baseUrlToken) this.baseUrl);
 }
+
+const listOfDurationToken = const OpaqueToken<List<Duration>>('listOfDuration');
+
+@Component(
+  selector: 'contains-child-component',
+  template: r'''
+    <div *ngIf="someValue">
+      <div *ngIf="someValue">
+        <injects-typed-token-from-parent></injects-typed-token-from-parent>
+      </div>
+    </div>
+  ''',
+  directives: const [
+    InjectsTypedTokenFromSomeParent,
+    NgIf,
+  ],
+  visibility: Visibility.local,
+)
+class ContainsChildComponent {
+  bool someValue = true;
+}
+
+@Component(
+  selector: 'injects-typed-token-from-parent',
+  template: '',
+  visibility: Visibility.local,
+)
+class InjectsTypedTokenFromSomeParent {
+  final List<Duration> example;
+
+  InjectsTypedTokenFromSomeParent(@Inject(listOfDurationToken) this.example);
+}